backend/requirements.txt
django==4.1.3
pillow==9.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-taggit==3.0.0
django-recurrence==1.11.1
geocoder==1.38.1
geographiclib==2.0 
geopy==2.3.0
django-filter==22.1
django-cors-headers==3.13.0